change: add remaining the validations for hp
Previously there where still some validations missing, so in this commit I have add them to healthcareProfessionalValidations
